+++ src/archive.h
@@ -17,7 +17,7 @@ namespace rsimpl
     public:
         struct frame_additional_data
         {
-            int timestamp = 0;
+            double timestamp = 0;
             int frame_number = 0;
             long long system_time = 0;
             int width = 0;
@@ -32,7 +32,7 @@ namespace rsimpl
 
             frame_additional_data(){};
 
-            frame_additional_data(int in_timestamp, int in_frame_number, long long in_system_time, int in_width, int in_height, int in_fps, int in_stride, float in_bpp, const rs_format in_format, rs_stream in_stream_type, int in_pad)
+            frame_additional_data(double in_timestamp, int in_frame_number, long long in_system_time, int in_width, int in_height, int in_fps, int in_stride, float in_bpp, const rs_format in_format, rs_stream in_stream_type, int in_pad)
                 :timestamp(in_timestamp),
                 frame_number(in_frame_number),
                 system_time(in_system_time),
@@ -82,8 +82,8 @@ namespace rsimpl
             ~frame() { on_release.reset(); }
 
             const byte* get_frame_data() const;
-            int get_frame_timestamp() const;
-            void set_timestamp(int new_ts) override { additional_data.timestamp = new_ts; }
+            double get_frame_timestamp() const;
+            void set_timestamp(double new_ts) override { additional_data.timestamp = new_ts; }
             int get_frame_number() const override;
             long long get_frame_system_time() const;
             int get_width()const;
@@ -147,7 +147,7 @@ namespace rsimpl
             }
 
             const byte* get_frame_data() const override;
-            int get_frame_timestamp() const override;
+            double get_frame_timestamp() const override;
             int get_frame_number() const override;
             long long get_frame_system_time() const override;
             int get_frame_width() const override;
@@ -176,7 +176,7 @@ namespace rsimpl
             }
 
             const byte * get_frame_data(rs_stream stream) const { return buffer[stream].get_frame_data(); }
-            int get_frame_timestamp(rs_stream stream) const { return buffer[stream].get_frame_timestamp(); }
+            double get_frame_timestamp(rs_stream stream) const { return buffer[stream].get_frame_timestamp(); }
             int get_frame_number(rs_stream stream) const { return buffer[stream].get_frame_number(); }
             long long get_frame_system_time(rs_stream stream) const { return buffer[stream].get_frame_system_time(); }
